refactor(NcSavingIndicatorIcon): migrate component to vue
