+++ src/components/NcSavingIndicatorIcon/index.ts
