+++ src/components/index.ts
@@ -71,7 +71,7 @@ export { default as NcProgressBar } from './NcProgressBar/index.js'
 export { default as NcRelatedResourcesPanel } from './NcRelatedResourcesPanel/index.js'
 export { default as NcRichContenteditable, NcAutoCompleteResult, NcMentionBubble } from './NcRichContenteditable/index.js'
 export { default as NcRichText } from './NcRichText/index.js'
-export { default as NcSavingIndicatorIcon } from './NcSavingIndicatorIcon/index.js'
+export { default as NcSavingIndicatorIcon } from './NcSavingIndicatorIcon/index.ts'
 export { default as NcSelect } from './NcSelect/index.js'
 export { default as NcSelectTags } from './NcSelectTags/index.js'
 export { default as NcSettingsInputText } from './NcSettingsInputText/index.js'
